requirements.txt
aiogram==3.3.0
aiohttp==3.9.1
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
aiosqlite==0.19.0
beautifulsoup4==4.12.2
lxml==5.1.0
yookassa==3.0.0
apscheduler==3.10.4
python-dotenv==1.0.0
telethon==1.34.0